Add files via upload
The style.css file is the visual designer of the registration form, transforming plain HTML into a clean and user-friendly interface. It defines the layout, colors, fonts, spacing, and overall aesthetic of the form — ensuring that users have a smooth and visually appealing experience. From centering the form on the page to styling buttons and input fields, this file brings the structure to life, reflecting the importance of design in frontend development. It helps convey that even the simplest functionality can feel polished with the right touch of styling.